// src/sdks/computation/computationV2ApiDocs.ts
export interface ReductionSceneReq {
  unitDescEn?: any;
  totalDescEn?: any;
  sceneNameEn?: any;
  sceneDescEn?: any;
  id?: number;

  orgId?: number;

  sceneDesc?: string;

  sceneName?: string;

  sceneType?: string;

  totalDesc?: string;

  totalEndValue?: number;

  totalLessenType?: '0' | '1';

  totalStartValue?: number;

  totalUnit?: '1' | '2' | '3' | '4' | '5' | '6';

  unitDenominatorUnit?: string;

  unitDesc?: string;

  unitEndValue?: number;

  unitLessenType?: '0' | '1';

  unitNumeratorUnit?: '1' | '2' | '3' | '4' | '5' | '6';

  unitStartValue?: number;
  languageSourceList?: any[];
}